reupload
So it's importing from a directory and thanks to your code it's creating a list : ship_list. I want it to use everywhere in my code but as it's under a class (MyMainApp, that I use to easily call the function in kv when I click on play with app.directory() ). So Under "Thirdwindow" I tried to equal this ship_list to Ships, which is used in most of my code. I also tried everywhere else but it doesn't work, I still don't have any ship ...